+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to the "neo4j-vscode" extension will be documented in this f
 ## [0.1.9]
 
 - Added support for Cypher code blocks in Markdown files
+- Automatically selects the text within the code block when running a query if no selection is made
+
+```cypher
+match (n) return n limit 2
+```
 
 [Reference](https://stackoverflow.com/a/76239666/3876654)